pushing changes with correct latching logic for behaviors and active monitors. however, the fundamental approach is still flawed.
next: rework so it's ONLY the interactors doing specific integration. all EBs should be the same and initializable as a list.

also, currently the information is flowing on the wrong "level".

next step also: integrate node layer to produce one output "current active" behavior, of which there can only ever be one.